requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.3.4
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
httpx==0.27.0
google-cloud-storage==2.18.2
python-dotenv==1.0.1
openai==1.54.1
